0.7.8 release

+++ example/add.c
@@ -2,11 +2,9 @@
 
 extern "C"
 {
-
 	EMSCRIPTEN_KEEPALIVE
-	int add(int a, int b)
+	float add(float a, float b)
 	{
 		return a + b;
 	}
-
 }
